photo-resizing/requirements.txt
Pillow==9.2.0
win32-setctime==1.1.0